requirements.txt
discord.py
pip~=21.0.1
attrs~=20.3.0
idna~=2.10
multidict~=5.1.0
yarl~=1.6.3
aiohttp~=3.7.4.post0
chardet~=3.0.4
setuptools~=49.2.1